make base abstract models available - wip
