+++ sl_tests/models.py
@@ -1,7 +1,5 @@
 from django.contrib.gis.db import models
 
-from simple_locations.base_area import AbstractBaseArea
-
 
 class WgsPoint(models.Model):
     position = models.PointField(srid=4326)
@@ -9,7 +7,3 @@ class WgsPoint(models.Model):
 
 class MercatorPoint(models.Model):
     position = models.PointField(srid=3857)
-
-class ConfiguredAreaModel(AbstractBaseArea):
-    extra_field = models.CharField(max_length=30)
-    extra_field2 = models.CharField(max_length=30, default='')
